Downgrade java version to 8
I have given this a deep thought, and after much consideration, I'll
be downgrading the Java version for two main reasons, my target audience,
and NachoSpigot's current state. You can read my reasons below.

Target audience; My main target audiences, believe it or not, is
mainly focused on the bridging aspect, and according to my [source](https://bstats.org/plugin/bukkit/SpeedBridge/12679), 59% of the
servers use 1.8.8, so that's that.

NachoSpigot's current state; Regardless of the point that I made
above, I always strongly recommended my users that use anything lower than
1.12.2 switch to NachoSpigot, a fork of TachoSpigot which supports Java
versions from 8 to the latest (17 as of now).

But now that they're no longer maintaining the project, I have no
choice but either to recommend something that is dead, go for alternatives or
simply downgrade the Java version to 8, and be over it.